Updated mksurfdata tools for new LUT
Updated the mksurfdat Makefile.common to fix the LDFLAGS line.
Changed a tolerance value in mksurfdat.F90 for near-zero machine error
to account for pft pct being rounded to 1e-7 rather than integer percent.
Updated the default namelist values for mksrf_fvegtyp to get the
correct files and to include all potential CMIP6 RCPs.
Added an explicit type conversion in updateannuallanduse_v2.c.

 Changes to be committed:
	modified:   components/clm/bld/namelist_files/namelist_defaults_clm4_5_tools.xml
	modified:   components/clm/tools/clm4_5/land_use_translator/updateannuallanduse_v2.c
	modified:   components/clm/tools/clm4_5/mksurfdata_map/src/Makefile.common
	modified:   components/clm/tools/clm4_5/mksurfdata_map/src/mksurfdat.F90

+++ components/clm/tools/clm4_5/land_use_translator/updateannuallanduse_v2.c
@@ -5914,7 +5914,7 @@ updateannuallanduse_main(double glmo[][GLMONFLDS], double plodata[][PLONFLDS], i
    char inbase[1000];
    memset(inbase, '\0', sizeof(inbase));
    char *extaddr = strrchr(usesfname, '.');
-   int numkeep = extaddr - usesfname;
+   int numkeep = (int) (extaddr - usesfname);
    strncpy(inbase, usesfname, numkeep);
    sprintf(out_land_filebase, "LUT_%s", inbase);
